+++ agate/utils.py
@@ -10,6 +10,7 @@ from collections import OrderedDict, Sequence
 from functools import wraps
 import string
 import warnings
+from slugify import Slugify, UniqueSlugify
 
 try:
     from cdecimal import Decimal, ROUND_FLOOR, ROUND_CEILING, getcontext
@@ -286,3 +287,22 @@ def issequence(obj):
     :class:`.Sequence` that is not also a string.
     """
     return isinstance(obj, Sequence) and not isinstance(obj, six.string_types)
+
+
+def slugify(values, ensure_unique=False, **kwargs):
+    """
+    Given a sequence of strings, returns a standardized version of the sequence.
+    If ``ensure_unique`` is True, any duplicate strings will be appended with
+    a unique identifier. Any kwargs will be passed to the Slugify or UniqueSlugify
+    class constructor (https://github.com/dimka665/awesome-slugify).
+    """
+    # Default to all lowercase
+    slug_args = {'to_lower': True}
+    slug_args.update(kwargs)
+
+    if ensure_unique:
+        slugify = UniqueSlugify(**slug_args)
+    else:
+        slugify = Slugify(**slug_args)
+
+    return tuple(slugify(value) for value in values)